requirements.txt
scipy==1.15.3
torchinfo==1.8.0
opencv_python==4.11.0.86
h5py==3.13.0
natsort==8.4.0
scikit-image==0.25.2
scikit-learn==1.6.1
torchsummary==1.5.1
kornia-rs==0.1.9
kornia==0.8.1